chore: 일정 추가 관련 model 위치 변경 및 분리
