+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/model/EventNotification.kt
@@ -1,0 +1,11 @@
+package com.teameetmeet.meetmeet.presentation.model
+
+import com.teameetmeet.meetmeet.R
+
+enum class EventNotification(val minutes: Long, val stringResId: Int) {
+    NONE(0, R.string.notification_none),
+    TEN_MINUTES(10, R.string.notification_ten_minutes_before),
+    ONE_HOUR(60, R.string.notification_one_hour_before),
+    ONE_DAY(24 * 60, R.string.notification_one_day_before),
+    ONE_WEEK(7 * 24 * 60, R.string.notification_one_week_before)
+}
